+++ cdc/tcp_dialer.go
@@ -15,6 +15,7 @@ import (
 
 const connectTimeout = 30 * time.Second
 
+// Connection is a subset of net.Connection to allow testing
 //go:generate counterfeiter -o ../mocks/connection.go --fake-name Connection . Connection
 type Connection interface {
 	Read(b []byte) (n int, err error)
@@ -22,10 +23,12 @@ type Connection interface {
 	Close() error
 }
 
+// TcpDialer opens a TCP connection to the given address
 type TcpDialer struct {
 	Address string
 }
 
+// Dial to the target with context and timeout
 func (t *TcpDialer) Dial(ctx context.Context) (Connection, error) {
 	glog.V(2).Infof("connect to %s", t.Address)
 	dialer := net.Dialer{