+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
+## 1.3.0
+
+- Add retry on connection error
+- Resume from last GTID on restart
+
+## 1.2.0
+
+- Retry read for JSON format
+
 ## 1.1.0
 
 - Allow configure UUID for setup streaming